Begin with persisting results

+++ src/SkiAnalyze.Core/PisteAggregate/PisteNode.cs
@@ -5,8 +5,8 @@ namespace SkiAnalyze.Core.PisteAggregate;
 public class PisteNode : BaseEntity<long>
 {
     public long OsmId { get; set; }
-    public float Longitude { get; set; }
-    public float Latitude { get; set; }
+    public double Longitude { get; set; }
+    public double Latitude { get; set; }
     public long PisteId { get; set; }
     public Piste? Piste { get; set; }
 }
